html/semantics/scripting-1/the-script-element/json-module/json-module-service-worker-test.https.tentative.html

<!doctype html>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
    promise_test(async (test) => {
        const reg = await navigator.serviceWorker.register('./serviceworker.js', { type: 'module' })
            .then((reg) => {
                //registration successful
                test.step_func_done(function () {
                    assert_not_equals(reg.installing, undefined);
                });
            });
    }, "Javascript importing JSON Module should load within the context of a service worker");

    promise_test(async (test) => {
        const reg = await navigator.serviceWorker.register('./module2.json', { type: 'module' })
            .then((reg) => {
                //registration successful
                test.step_func_done(function () {
                    assert_not_equals(reg.installing, undefined);
                });
            });
    }, "JSON Modules should load within the context of a service worker");

    promise_test(async (test) => {
        const reg = await navigator.serviceWorker.register('./serviceworker-dynamic-import.js', { type: 'module' })
            .then((reg) => {
                //registration successful
                test.step_func_done(function () {
                    assert_not_equals(reg.installing, undefined);
                });
            });
    }, "JSON Module dynamic import should load within the context of a service worker");
</script>
